Move front-end deps into dev dependencies
The reasoning here is that in prod we will have a compiled and minified JS file
along with a similar CSS file. Since these are precompiled we don't actually
need any of the dev deps in that environment.

However it should be noted that 'devDependencies' is misleading, since those
deps are actually _build_ dependencies. I may rethink this strategy later, but
for now only server side deps make the cut for `dependencies`.